requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.31
requests>=2.31.0
ta>=0.11.0
vaderSentiment>=3.3.2
textblob>=0.17.1
xgboost>=2.0.0
scikit-learn>=1.3.0
statsmodels>=0.14.0
plotly>=5.17.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
